requirements.txt
boto3==1.34.100
botocore==1.34.107
mysql_connector_repackaged==0.3.1
python-dotenv==1.0.1
websocket_client==1.8.0
